Moving the linked list inside the graph classes
