+++ data_structures/graph_classes/linked_list/list_node.py
